backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
pydantic
pydantic-settings
livekit
livekit-agents
livekit-plugins-deepgram
livekit-plugins-google
livekit-plugins-cartesia
livekit-plugins-silero
livekit-plugins-turn-detector
pinecone-client
pinecone-plugin-assistant
python-multipart
Pillow
numpy
pinecone
livekit-agents[anam]